requirements.txt
Django==1.7.4
Unipath==1.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
django-user-agents==0.3.0
djangorestframework==3.0.4
gunicorn==19.2.0
psycopg2==2.5.4
static3==0.5.1
ua-parser==0.3.6
user-agents==0.3.2
wsgiref==0.1.2
